requirements.txt
python-dotenv~=0.21.0
selenium~=4.6.0
requests~=2.28.1
beautifulsoup4~=4.11.1